app/requirements.txt
boto3==1.17.49
botocore==1.20.49
jmespath==0.10.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
s3transfer==0.3.6
six==1.15.0
urllib3==1.26.4
